refactor: adjust meta tag head
