+++ src/pages/_document.tsx
@@ -2,10 +2,8 @@ import { ColorModeScript } from "@chakra-ui/react";
 import type { DocumentContext } from "next/document";
 import Document, { Html, Head, Main, NextScript } from "next/document";
 
-import { UMAMI_SRC, UMAMI_WEBSITE_ID } from "lib/constants/umami";
+import Meta from "lib/components/Meta";
 import customTheme from "lib/styles/theme";
-
-export const APP_NAME = "sznm.dev";
 
 class MyDocument extends Document {
   static getInitialProps(ctx: DocumentContext) {
@@ -25,34 +23,7 @@ class MyDocument extends Document {
             rel="stylesheet"
           />
 
-          <meta name="application-name" content={APP_NAME} />
-          <meta name="apple-mobile-web-app-capable" content="yes" />
-          <meta
-            name="apple-mobile-web-app-status-bar-style"
-            content="default"
-          />
-          <meta name="apple-mobile-web-app-title" content={APP_NAME} />
-          <meta name="format-detection" content="telephone=no" />
-          <meta name="mobile-web-app-capable" content="yes" />
-          <meta name="theme-color" content="#444444" />
-
-          {/* add your own app-icon */}
-          {/* <link
-            rel="apple-touch-icon"
-            sizes="180x180"
-            href="/icons/apple-touch-icon.png"
-          />
-          <link rel="icon" href="/app-icon.png" /> */}
-          <link rel="manifest" href="/manifest.json" />
-
-          {/* umami self-hosted analytics */}
-          <script
-            async
-            defer
-            data-website-id={UMAMI_WEBSITE_ID}
-            src={UMAMI_SRC}
-            data-domains="sznm.dev"
-          />
+          <Meta />
         </Head>
         <body>
           <ColorModeScript
